Convert tests from strategy to E2E
